Refactor cli and python interface into html2htpy
