--- htpy/utils/__init__.py
@@ -1,1 +1,0 @@
-from htpy.utils.html_to_htpy import *
